Move FILEIO structure definitions to common header files
